Address DCS review: restore existing plugins, relocate diffusion to Continuous
- Restore BioNetwork.{h,cpp} to original (no changes to the existing biochemical plugin)
- Restore test_simulator_runtime.cpp to original (do not modify the existing test file)
- Move DiffusionField plugin from data/DiffusionSimulation/ to data/Continuous/
- Add standalone DiffusionField integration tests (test_plugins_continuous_diffusion.cpp)
- Restore root README.md to the general GenESyS documentation
- Remove relatorio.pdf and video_entrega.mp4 from the source repository
