+++ source/tests/unit/CMakeLists.txt
@@ -277,6 +277,15 @@ genesys_add_unit_test(genesys_test_tools_diffusion_mol
     genesys_tools
 )
 
+# DCS: DiffusionField plugin integration (Part 2) — drives the plugin through the real kernel
+genesys_add_unit_test(genesys_test_plugins_continuous_diffusion
+    test_plugins_continuous_diffusion.cpp
+    genesys_tools
+    genesys_kernel_simulator_runtime
+    genesys_kernel_util
+    genesys_plugins_data
+)
+
 add_custom_target(genesys_kernel_unit_tests
     DEPENDS
         genesys_test_util
@@ -298,6 +307,7 @@ add_custom_target(genesys_kernel_unit_tests
         genesys_test_tools_simulation_results_dataset
         genesys_test_tools_ode_solver_factory
         genesys_test_tools_diffusion_mol
+        genesys_test_plugins_continuous_diffusion
 )
 
 if(GENESYS_BUILD_WEB_APPLICATION)
@@ -379,6 +389,7 @@ add_custom_target(genesys_kernel_unit_tests_run
     COMMAND $<TARGET_FILE:genesys_test_tools_simulation_results_dataset>
     COMMAND $<TARGET_FILE:genesys_test_tools_ode_solver_factory>
     COMMAND $<TARGET_FILE:genesys_test_tools_diffusion_mol>
+    COMMAND $<TARGET_FILE:genesys_test_plugins_continuous_diffusion>
     COMMAND $<TARGET_FILE:genesys_test_kernel_simulator_method_inventory>
     DEPENDS
         genesys_kernel_unit_tests
