+++ source/tests/unit/test_simulator_runtime.cpp
@@ -40,7 +40,6 @@
 #include "plugins/data/BiochemicalSimulation/BioParameter.h"
 #include "plugins/data/BiochemicalSimulation/BioReaction.h"
 #include "plugins/data/BiochemicalSimulation/BioSpecies.h"
-#include "plugins/data/DiffusionSimulation/DiffusionField.h"
 #include "plugins/data/BiochemicalSimulation/GroProgram.h"
 #include "plugins/data/BiochemicalSimulation/GeneticCircuitPart.h"
 #include "plugins/data/BiochemicalSimulation/GeneticCircuit.h"
@@ -8838,319 +8837,4 @@ TEST(SimulatorRuntimeTest, CppSerializerEmitsCurrentApiAndPropertySetters) {
     EXPECT_NE(generated.find("Delay_A->connectTo(Dispose_A, 0);"), std::string::npos);
     EXPECT_NE(generated.find("sim->setNumberOfReplications(7);"), std::string::npos);
     EXPECT_NE(generated.find("sim->setReplicationLength(42.000000, Util::TimeUnit::minute);"), std::string::npos);
-}
-
-
-// ===========================================================================
-//  DCS: BioNetwork ODE-solver-selection integration tests (Part 1)
-// ===========================================================================
-
-TEST(SimulatorRuntimeTest, BioNetworkDefaultsToRungeKutta4Solver) {
-    Simulator simulator;
-    Model* model = simulator.getModelManager()->newModel();
-    ASSERT_NE(model, nullptr);
-    BioNetworkProbe network(model, "DefaultSolverNetwork");
-    EXPECT_EQ(network.getOdeSolver(), "RungeKutta4");
-}
-
-TEST(SimulatorRuntimeTest, BioNetworkSimulatesWithDormandPrinceSolver) {
-    Simulator simulator;
-    Model* model = simulator.getModelManager()->newModel();
-    ASSERT_NE(model, nullptr);
-
-    BioSpeciesProbe a(model, "A");
-    a.setInitialAmount(10.0);
-    a.setAmount(10.0);
-    BioSpeciesProbe b(model, "B");
-    b.setInitialAmount(0.0);
-    b.setAmount(0.0);
-
-    BioReactionProbe reaction(model, "A_to_B");
-    reaction.addReactant("A", 1.0);
-    reaction.addProduct("B", 1.0);
-    reaction.setRateConstant(0.1);
-    reaction.setReverseRateConstant(0.05);
-    reaction.setReversible(true);
-
-    std::string errorMessage;
-    ASSERT_TRUE(reaction.CheckProbe(errorMessage)) << errorMessage;
-
-    BioNetworkProbe network(model, "DormandPrinceNetwork");
-    network.setOdeSolver("DormandPrince54");
-    ASSERT_EQ(network.getOdeSolver(), "DormandPrince54");
-    // Coarser step than the RK4 reference test (0.05 vs 0.001): the adaptive
-    // solver should still reach the exact equilibrium.
-    ASSERT_TRUE(network.simulate(0.0, 1.0, 0.05, errorMessage)) << errorMessage;
-
-    const double total = 10.0;
-    const double forwardRate = 0.1;
-    const double reverseRate = 0.05;
-    const double equilibriumA = reverseRate * total / (forwardRate + reverseRate);
-    const double expectedA = equilibriumA + (10.0 - equilibriumA) * std::exp(-(forwardRate + reverseRate));
-    EXPECT_EQ(network.getLastStatus(), "Completed");
-    EXPECT_NEAR(a.getAmount(), expectedA, 1e-4);
-    EXPECT_NEAR(b.getAmount(), total - expectedA, 1e-4);
-}
-
-TEST(SimulatorRuntimeTest, BioNetworkSolversAgreeOnSameModel) {
-    Simulator simulator;
-    Model* model = simulator.getModelManager()->newModel();
-    ASSERT_NE(model, nullptr);
-
-    BioSpeciesProbe a(model, "A");
-    a.setInitialAmount(10.0);
-    a.setAmount(10.0);
-    BioSpeciesProbe b(model, "B");
-    b.setInitialAmount(0.0);
-    b.setAmount(0.0);
-    BioReactionProbe reaction(model, "A_to_B");
-    reaction.addReactant("A", 1.0);
-    reaction.addProduct("B", 1.0);
-    reaction.setRateConstant(0.3);
-
-    std::string errorMessage;
-    ASSERT_TRUE(reaction.CheckProbe(errorMessage)) << errorMessage;
-
-    BioNetworkProbe rkNet(model, "RkNet");
-    rkNet.setOdeSolver("RungeKutta4");
-    ASSERT_TRUE(rkNet.simulate(0.0, 2.0, 0.01, errorMessage)) << errorMessage;
-    const double rkA = a.getAmount();
-
-    a.setAmount(10.0);
-    b.setAmount(0.0);
-    BioNetworkProbe dpNet(model, "DpNet");
-    dpNet.setOdeSolver("DormandPrince54");
-    ASSERT_TRUE(dpNet.simulate(0.0, 2.0, 0.01, errorMessage)) << errorMessage;
-    const double dpA = a.getAmount();
-
-    // Both methods integrate the same ODE; they must agree closely.
-    EXPECT_NEAR(rkA, dpA, 1e-5);
-    // And both must match the closed-form first-order decay A(t)=A0 e^{-k t}.
-    EXPECT_NEAR(dpA, 10.0 * std::exp(-0.3 * 2.0), 1e-4);
-}
-
-TEST(SimulatorRuntimeTest, BioNetworkOdeSolverFieldRoundTripsThroughPersistence) {
-    Simulator simulator;
-    Model* model = simulator.getModelManager()->newModel();
-    ASSERT_NE(model, nullptr);
-
-    BioNetworkProbe source(model, "PersistSolver");
-    source.setOdeSolver("DormandPrince54");
-
-    FakeModelPersistenceRuntime persistence;
-    PersistenceRecord fields(persistence);
-    source.SaveInstanceProbe(&fields, true);
-
-    BioNetworkProbe loaded(model, "LoadedPersistSolver");
-    ASSERT_TRUE(loaded.LoadInstanceProbe(&fields));
-    EXPECT_EQ(loaded.getOdeSolver(), "DormandPrince54");
-}
-
-TEST(SimulatorRuntimeTest, BioNetworkRejectsUnknownOdeSolverKey) {
-    Simulator simulator;
-    Model* model = simulator.getModelManager()->newModel();
-    ASSERT_NE(model, nullptr);
-
-    BioSpeciesProbe a(model, "A");
-    a.setInitialAmount(1.0);
-    a.setAmount(1.0);
-    BioReactionProbe reaction(model, "Adecay");
-    reaction.addReactant("A", 1.0);
-    reaction.setRateConstant(0.1);
-
-    BioNetworkProbe network(model, "BadSolverNetwork");
-    network.setOdeSolver("NotARealSolver");
-
-    std::string errorMessage;
-    EXPECT_FALSE(network.CheckProbe(errorMessage));
-    EXPECT_NE(errorMessage.find("unknown ODE solver"), std::string::npos);
-}
-
-
-// ===========================================================================
-//  DCS: DiffusionField N-D Method-of-Lines integration tests (Part 2)
-// ===========================================================================
-
-class DiffusionFieldProbe : public DiffusionField {
-public:
-    DiffusionFieldProbe(Model* model, const std::string& name = "") : DiffusionField(model, name) {}
-    bool CheckProbe(std::string& errorMessage) { return _check(errorMessage); }
-    void SaveInstanceProbe(PersistenceRecord* fields, bool saveDefaultValues = false) { _saveInstance(fields, saveDefaultValues); }
-    bool LoadInstanceProbe(PersistenceRecord* fields) { return _loadInstance(fields); }
-};
-
-namespace {
-// Decay rate of a sine mode on a uniform N-D grid (from the discrete Laplacian):
-// lambda = D * N * (4/h^2) * sin^2(mode*pi / (2*(points-1))).
-double discreteSineDecayRate(unsigned int dims, unsigned int points, double length,
-                             double D, unsigned int mode) {
-    const double h = length / (points - 1);
-    const double s = std::sin(mode * M_PI / (2.0 * (points - 1)));
-    return D * dims * (4.0 / (h * h)) * s * s;
-}
-} // namespace
-
-TEST(SimulatorRuntimeTest, DiffusionFieldDefaultsToRungeKutta4) {
-    Simulator simulator;
-    Model* model = simulator.getModelManager()->newModel();
-    ASSERT_NE(model, nullptr);
-    DiffusionFieldProbe field(model, "DiffusionDefaultSolver");
-    EXPECT_EQ(field.getOdeSolver(), "RungeKutta4");
-    EXPECT_EQ(field.getDimensions(), 2u);
-}
-
-TEST(SimulatorRuntimeTest, DiffusionFieldMatchesAnalyticalEigenmodeDecay) {
-    Simulator simulator;
-    Model* model = simulator.getModelManager()->newModel();
-    ASSERT_NE(model, nullptr);
-
-    DiffusionFieldProbe field(model, "DiffusionEigenmode");
-    field.setDimensions(1u);
-    field.setPointsPerDimension(41u);
-    field.setDomainLength(1.0);
-    field.setDiffusionCoefficient(0.2);
-    field.setBoundaryCondition("Dirichlet");
-    field.setInitialCondition("SineMode");
-    field.setInitialParameter(2.0); // mode 2
-    field.setOdeSolver("DormandPrince54");
-
-    std::string errorMessage;
-    ASSERT_TRUE(field.simulate(0.0, 0.5, 0.05, errorMessage)) << errorMessage;
-    EXPECT_EQ(field.getLastStatus(), "Completed");
-
-    // Sample an interior node and compare to v0 * exp(-lambda T).
-    const double h = 1.0 / 40.0;
-    const unsigned int m = 13;
-    const double v0 = std::sin(2.0 * M_PI * (m * h) / 1.0);
-    const double lambda = discreteSineDecayRate(1u, 41u, 1.0, 0.2, 2u);
-    const double expected = v0 * std::exp(-lambda * 0.5);
-    EXPECT_NEAR(field.getFieldValue({m}), expected, 1e-4);
-}
-
-TEST(SimulatorRuntimeTest, DiffusionFieldSolversAgree) {
-    Simulator simulator;
-    Model* model = simulator.getModelManager()->newModel();
-    ASSERT_NE(model, nullptr);
-
-    auto run = [&](const std::string& solver, double sampleNode) {
-        DiffusionFieldProbe field(model, "DiffusionAgree_" + solver);
-        field.setDimensions(1u);
-        field.setPointsPerDimension(41u);
-        field.setDomainLength(1.0);
-        field.setDiffusionCoefficient(0.05);
-        field.setInitialCondition("SineMode");
-        field.setInitialParameter(3.0);
-        field.setOdeSolver(solver);
-        std::string err;
-        EXPECT_TRUE(field.simulate(0.0, 0.2, 0.005, err)) << err;
-        return field.getFieldValue({static_cast<unsigned int>(sampleNode)});
-    };
-    const double rk = run("RungeKutta4", 20);
-    const double dp = run("DormandPrince54", 20);
-    EXPECT_NEAR(rk, dp, 1e-4);
-}
-
-TEST(SimulatorRuntimeTest, DiffusionFieldNeumannConservesMass) {
-    Simulator simulator;
-    Model* model = simulator.getModelManager()->newModel();
-    ASSERT_NE(model, nullptr);
-
-    DiffusionFieldProbe field(model, "DiffusionNeumannMass");
-    field.setDimensions(2u);
-    field.setPointsPerDimension(31u);
-    field.setDomainLength(3.0);
-    field.setDiffusionCoefficient(0.3);
-    field.setBoundaryCondition("Neumann");
-    field.setInitialCondition("Gaussian");
-    field.setInitialParameter(0.3); // sigma
-    field.setOdeSolver("DormandPrince54");
-
-    std::string err;
-
-    // First run: initialize the field and obtain the reference mass.
-    ASSERT_TRUE(field.simulate(0.0, 0.1, 0.05, err)) << err;
-    const double massStart = field.getTotalMass();
-
-    // Continue the simulation and verify mass conservation.
-    ASSERT_TRUE(field.simulate(0.1, 0.5, 0.05, err)) << err;
-
-    EXPECT_NEAR(field.getTotalMass(), massStart, 1e-6);
-    EXPECT_LT(field.getMaxValue(), 1.0); // Gaussian spreads, peak drops
-}
-
-TEST(SimulatorRuntimeTest, DiffusionFieldRunsInThreeDimensions) {
-    Simulator simulator;
-    Model* model = simulator.getModelManager()->newModel();
-    ASSERT_NE(model, nullptr);
-
-    DiffusionFieldProbe field(model, "Diffusion3D");
-    field.setDimensions(3u);          // works for any N
-    field.setPointsPerDimension(13u);
-    field.setDomainLength(1.0);
-    field.setDiffusionCoefficient(0.1);
-    field.setInitialCondition("SineMode");
-    field.setInitialParameter(1.0);
-    field.setOdeSolver("DormandPrince54");
-
-    std::string err;
-    ASSERT_TRUE(field.simulate(0.0, 0.3, 0.05, err)) << err;
-    EXPECT_EQ(field.getField().size(), 13u * 13u * 13u);
-
-    const unsigned int c = 6; // center node
-    const double h = 1.0 / 12.0;
-    const double v0 = std::pow(std::sin(M_PI * (c * h)), 3.0);
-    const double lambda = discreteSineDecayRate(3u, 13u, 1.0, 0.1, 1u);
-    EXPECT_NEAR(field.getFieldValue({c, c, c}), v0 * std::exp(-lambda * 0.3), 1e-3);
-}
-
-TEST(SimulatorRuntimeTest, DiffusionFieldConfigRoundTripsThroughPersistence) {
-    Simulator simulator;
-    Model* model = simulator.getModelManager()->newModel();
-    ASSERT_NE(model, nullptr);
-
-    DiffusionFieldProbe source(model, "DiffusionPersistSource");
-    source.setDimensions(3u);
-    source.setPointsPerDimension(17u);
-    source.setDomainLength(2.5);
-    source.setDiffusionCoefficient(0.42);
-    source.setBoundaryCondition("Neumann");
-    source.setInitialCondition("Gaussian");
-    source.setInitialParameter(0.2);
-    source.setOdeSolver("DormandPrince54");
-
-    FakeModelPersistenceRuntime persistence;
-    PersistenceRecord fields(persistence);
-    source.SaveInstanceProbe(&fields, true);
-
-    DiffusionFieldProbe loaded(model, "DiffusionPersistLoaded");
-    ASSERT_TRUE(loaded.LoadInstanceProbe(&fields));
-    EXPECT_EQ(loaded.getDimensions(), 3u);
-    EXPECT_EQ(loaded.getPointsPerDimension(), 17u);
-    EXPECT_DOUBLE_EQ(loaded.getDomainLength(), 2.5);
-    EXPECT_DOUBLE_EQ(loaded.getDiffusionCoefficient(), 0.42);
-    EXPECT_EQ(loaded.getBoundaryCondition(), "Neumann");
-    EXPECT_EQ(loaded.getInitialCondition(), "Gaussian");
-    EXPECT_EQ(loaded.getOdeSolver(), "DormandPrince54");
-}
-
-TEST(SimulatorRuntimeTest, DiffusionFieldRejectsInvalidSolverAndConfig) {
-    Simulator simulator;
-    Model* model = simulator.getModelManager()->newModel();
-    ASSERT_NE(model, nullptr);
-
-    DiffusionFieldProbe badSolver(model, "DiffusionBadSolver");
-    badSolver.setOdeSolver("NotARealSolver");
-    std::string e1;
-    EXPECT_FALSE(badSolver.CheckProbe(e1));
-    EXPECT_NE(e1.find("unknown ODE solver"), std::string::npos);
-
-    DiffusionFieldProbe badGrid(model, "DiffusionBadGrid");
-    badGrid.setPointsPerDimension(2u); // below the minimum of 3
-    std::string e2;
-    EXPECT_FALSE(badGrid.CheckProbe(e2));
-
-    DiffusionFieldProbe badBoundary(model, "DiffusionBadBoundary");
-    badBoundary.setBoundaryCondition("Periodic"); // unsupported
-    std::string e3;
-    EXPECT_FALSE(badBoundary.CheckProbe(e3));
 }
